handle connection failures on the builds page
for browsers that support the Page Visibility API, it will automatically
reconnect once per minute in the case of a hard connection error (one
where the browser does not decide to automatically reconnect to the
event source). if the page is not visible, reconnecting is delayed until
the page becomes visible.

when there is not an active connection or an in-progress connection
attempt, a button is displayed to allow the user to reconnect.